A lot of changes
Got rid of donate button within the window. Added a menu bar with
traditional options such as Save, Close, About. Most of these options
have placeholder functions as not all can apply to the current version
of our algorithm. Donation button was transformed as one of these menu
options. Slightly changed size of screen for better visual appeal.
Covered corner case of an article field not being entered. The program
will tell the user that a field is blank.